+++ .eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  'extends': 'ryansobol/es5',
+  'rules': {
+    // For consistency with keys that require quotes
+    'quote-props': 0
+  }
+};