+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All changes to this project are documented in this file. This project adheres to
 - Default to ES5 rules
 - Update CircleCI and NPM badges
 - Update `node` dependency to ^5.1.0
+- Update `.eslintrc.js` to extend `ryansobol/es5`
 
 ## [5.9.0] - 2016-01-19
 ### Changed